requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
python-dotenv
streamlit
requests
pypdf
tavily-python
python-multipart
pandas  
 